% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewaymanagementapi_package.R
\docType{package}
\name{paws.apigatewaymanagementapi-package}
\alias{paws.apigatewaymanagementapi}
\alias{paws.apigatewaymanagementapi-package}
\title{paws.apigatewaymanagementapi: AmazonApiGatewayManagementApi}
\description{
The Amazon API Gateway Management API allows you to
    directly manage runtime aspects of your deployed APIs. To use it, you
    must explicitly set the SDK\'s endpoint to point to the endpoint of
    your deployed API. The endpoint will be of the form
    https://{api-id}.execute-api.{region}.amazonaws.com/{stage}, or will
    be the endpoint corresponding to your API\'s custom domain and base
    path, if applicable.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
