% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigatewaymanagementapi_operations.R
\name{post_to_connection}
\alias{post_to_connection}
\title{Sends the provided data to the specified connection}
\usage{
post_to_connection(Data, ConnectionId)
}
\arguments{
\item{Data}{[required] The data to be sent to the client specified by its connection id.}

\item{ConnectionId}{[required] The identifier of the connection that a specific client is using.}
}
\description{
Sends the provided data to the specified connection.
}
\section{Accepted Parameters}{
\preformatted{post_to_connection(
  Data = raw,
  ConnectionId = "string"
)
}
}

